services/api-server/requirements.txt
click==8.1.3
Flask==2.2.3
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
PyYAML==6.0
RPi.GPIO==0.7.1
Werkzeug==2.2.3
zipp==3.14.0
